PlannerConfiguration types
