--- SharePointFramework/ProjectExtensions/src/tasks/PlannerConfiguration/IPlannerPlan.ts
@@ -1,4 +1,0 @@
-export interface IPlannerPlan {
-    id: string;
-    title: string;
-}
